Refresh d3d8to9 code with v1.6.0
- Minor log update
- Update Utils::LoadLibrary logic
- Refresh d3d8to9 code
- Use Use '.find()' for AddressCache lookup to fix an issue where using
operator[] on a map would create a key with value 0, thus returned
pAddress8 is nullptr and it's put in the map that way.
- Added SetFvF to initialize vertex declaration to fix an issue with Max
Payne.
- Removed SetSoftwareVertexProcessing() to fix an issue with Haegemonia:
Legions of Iron.
- Modified the code that translates the vertex declaration so that if
the Usage is blendindices it will set the Method to Default.  This fixes
an error in CreateVertexDeclaration found in True Crime New York City.
- Updated the pixel shader translation to ignore co-issued statements to
resolve an issue found in True Crime New York City.
- Updated the pixel shader translation to ensure that double swizzles
are not added when both the modified and modifier registers both have
swizzles.  This fixes an issue found in True Crime New York City.