Update API used to disable High DPI scaling
The previous API would only work if you called it early enough, which
was not alwasy possible.  It seemed to almost never work.  The new API
is more relyable and should always work on Windows 8.1/10, assuming the
application does not call this API, in which case we don't need to.